requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
langchain>=0.2.27
langchain-google-genai>=0.0.10
langchain-community>=0.0.11
langchain-text-splitters>=0.0.7
faiss-cpu>=1.7.4

# pip install -U streamlit python-dotenv langchain langchain-google-genai langchain-community langchain-text-splitters faiss-cpu
